endpoints added get(all books/by {id}/{title}/{publish_year}/{isbn}/{authorid}/available books/unavailable books/by author{id}) post(add book) & delete by {id}

+++ src/main/java/pl/wsb/lesinskibartosz/LibrarySystem/Repository/LibraryUserRepository.java
@@ -3,5 +3,5 @@ package pl.wsb.lesinskibartosz.LibrarySystem.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.wsb.lesinskibartosz.LibrarySystem.model.LibraryUser;
 
-public interface LibraryUserRepository extends JpaRepository<LibraryUser, Integer> {
+public interface LibraryUserRepository extends JpaRepository<LibraryUser, Long> {
 }
